Major changes to radolan_binaries_to_dataframe concerning missing data
The handling of binary files, which could not be read in has been changed.
Up to now, the processing of an entire month has been skipped when any file of the respective month could not be imported.
From now on, the processing continues, the corrupted files are skipped and the respective intervals are filled with NoData values.
Moreover, to inform the user and obtain more information on the errors, a text file containing the names of all skipped files and the corresponding error messages is written.

raw.create_idraster_and_process_radolan_data and process_radolan_data:
The creation of a text file containing a list of all skipped months due to processing errors has been removed
in accordance with the changes to radproc.raw.radolan_binaries_to_dataframe.
As the latter function is executed once for each month, one textfile is written for every month in which any processing errors occurred.